requirements.txt
chromadb>=0.5.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=11.0.0
mcp>=1.0.0
httpx>=0.28.0
